Relocate spotguide package
