+++ src/spotguide/cicdconfig.go
