+++ src/spotguide/scm/scm.go
