+++ src/spotguide/spotguide_test.go
